Promote on Disconnect (#4)
* hyparview: PromotePassiveBut on Disconnect (with printf)

* hyparview_test: looking for asymmetries

* simulation/client: keep each client's message history

* simulation/world_plot: print one lost client's history

* go.mod: pretty

* failure: Send interface return the specific struct pointer

go has a feature where using the Message interface in the function
signature but returning a nil struct will make the interface not nil,
it's `type` slot will be filled while the `value` slot is nil, and the
nil comparison will fail. The fix appears to be returning an explicit
nil in the function that knows about the struct type but returns the
interface type.

This function is in the client code, so I'm just going to use the
specific type all the way down to the interface. Otherwise, every
implementing client will have to endure at least one inexplicable bug,
where failure recovery doesn't work right, but also doesn't fail.

* hyparview: ForwardJoin uses From() to prune, not Join

* simulation/world: we don't need shouldFail

* README: fix link to peer sample paper

* hyparview: unused import

* hyparview_test: change log to test